livecode

+++ db/seeds.rb
@@ -5,17 +5,44 @@
 #
 #   movies = Movie.create([{ name: "Star Wars" }, { name: "Lord of the Rings" }])
 #   Character.create(name: "Luke", movie: movies.first)
+Garden.destroy_all if Rails.env.development?
+Tag.destroy_all
 
-puts "Destroying gardens.."
-Garden.destroy_all
-
-puts "Creating gardens.."
-Garden.create!(
+little = Garden.create!(
   name: "My Little Garden",
   banner_url: "https://raw.githubusercontent.com/lewagon/fullstack-images/master/rails/parks-and-plants/garden_1.jpg"
 )
 
-Garden.create!(
+other = Garden.create!(
   name: "My Other Garden",
   banner_url: "https://raw.githubusercontent.com/lewagon/fullstack-images/master/rails/parks-and-plants/garden_2.jpg"
 )
+
+Plant.create!(
+  name: "Monstera",
+  image_url: "https://raw.githubusercontent.com/lewagon/fullstack-images/master/rails/parks-and-plants/plants/monstera.jpg",
+  garden: little
+)
+
+Plant.create!(
+  name: "Philo",
+  image_url: "https://raw.githubusercontent.com/lewagon/fullstack-images/master/rails/parks-and-plants/plants/philo.jpg",
+  garden: little
+)
+
+Plant.create!(
+  name: "Dieff",
+  image_url: "https://raw.githubusercontent.com/lewagon/fullstack-images/master/rails/parks-and-plants/plants/dieffenbachia.jpg",
+  garden: other
+)
+
+Tag.create!(name: "Toxic")
+Tag.create!(name: "Exotic")
+Tag.create!(name: "Wild")
+Tag.create!(name: "Smelly")
+
+# Pseudocode:
+# 1. Generate Tag model
+# 2. Generate plant_tag model
+# 3. Generate the associations between models
+# 4. Migrate & test & create seeds for tags
